Introduced query-parser mw :recycle:
