+++ server/types/framework.ts
@@ -1,11 +1,12 @@
+import { Pagination, ParsedQuery } from '../shared/query-parser/types';
 import { File } from 'multer';
-import { Pagination } from '../middleware/pagination';
 import User from '../user/model';
 
 declare module 'express' {
   interface Request {
     user: User;
     pagination: Pagination;
+    parsedQuery: ParsedQuery
     file: File;
   }
 }
